Release v2.2.0: Memory optimizations, parallel scan, and type system
Added:
- True parallel scan (Blelloch algorithm) for O(log n) depth
- Gradient checkpointing for FlowingNEXUS memory efficiency
- FlowingContinualLearner with curriculum learning
- Comprehensive type system (nexus/core/types.py)
- EMA update callback in NEXUSTrainer
- Integration tests for full pipeline

Fixed:
- Device handling across all modules
- Duplicate server endpoints
- Memory leaks in trajectory storage

Changed:
- FlowingConfig: Added gradient_checkpointing options
- TrainingConfig: Added EMA configuration
- Documentation updates in copilot-instructions.md

+++ nexus/core/energy.py
@@ -173,7 +173,7 @@ class IterativeRefinement(nn.Module):
         energies = []
         
         max_iter = n_iterations or self.config.max_iterations
-        actual_iterations = torch.zeros(batch, device=device)
+        actual_iterations = torch.zeros(batch, device=device, dtype=x.dtype)
         converged = torch.zeros(batch, dtype=torch.bool, device=device)
         
         for i in range(max_iter):
